+++ src/index.ts
@@ -12,18 +12,27 @@ export function yaclil(rootCommand: Command, cliName: string, options?: YaclilOp
     options ??= {};
     const argv = options.customArgv ?? process.argv;
     // call the parser
-    const { parsed, helpCalled } = parseCommand({
+    const parser = parseCommand({
         rootCommand,
         rootCommandName: cliName,
         initOptions: options,
         // remove the first two arguments
         argv: argv.slice(2)
     });
+    const helpCalled = parser.helpCalled;
+    const parsed = parser.parsedObject;
     // get the configured value of include help feature, or `true` by default
     const helpConfigValue = parsed.command.helpOption ?? options.helpOptions ?? true;
     if (helpCalled && helpConfigValue) {
         // return the help string
         console.log(renderHelp(parsed));
+    } else if (parser.standaloneOptionCalled) {
+        // if standalone option called
+        const parsedForStandalone = parser.parsedStandaloneOption;
+        const handler = parsedForStandalone.parsedOption.option.standaloneHandler;
+        if (!handler) throw new Error('Impossible error with standalone options');
+        // execute standalone option
+        handler(parsedForStandalone);
     } else {
         // if this command is not callable
         if (!parsed.command.handler) {